+++ tests/testthat/test-prepInputsStandAgeMap.R
@@ -32,7 +32,7 @@ testthat::test_that("prepInputsStandAgeMap work", {
   out4 <- prepInputsStandAgeMap(dataSource = "SCANFI",
                                 dataYear = 2020, studyArea = goodPoly,
                                 rasterToMatch = goodRas, destinationPath = td,
-                                overwrite = TRUE, fireURL = NULL)
+                                overwrite = TRUE)
   expect_true(length(attr(out4, "imputedPix")) != length(attr(out, "imputedPix")))