+++ NEWS.md
@@ -1,5 +1,6 @@
 # LandR (development version)
 
+* SCANFI and 2020 now default data source and year for stand age and biomass functions
 * drop support for R 4.2 due to changes in dependency packages;
 * remove deprecated package `crayon` in favour of `cli` instead;
 * `dropTerm` now can deal with random effects better (#105);